+++ test/background_migrations/background_migrations.rb
@@ -24,6 +24,8 @@ module BackgroundMigrations
   end
 
   class MakeAllNonAdmins < OnlineMigrations::BackgroundMigration
+    def initialize(*_dummy_args) end
+
     def relation
       User.all
     end